pycurl - a python interface to libcurl. http://pycurl.sf.net>
currently pycurl requires that libcurl of equal or greater version number
is installed (which makes sense since this is just an interface).
unfortunately dp doesn't have any versioning dependencies yet. until
then, just force the install of the curl port with a bad regex. this
works the first time but little more.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@1282 d073be05-634f-4543-b044-5fe20cf6d1d6