Submitted by:   jmpp@ Reviewed by:    jmpp@, olegb@, mww@ Approved by:    jmpp@
Finally, the aqua variant of the aMSN port! Just the initial checkin, as it stands the variant could use
a *lot* of improvement, but it works so I'm committing it because I've been holding it back for too long
already. So basically if anyone reading this wants to take a stab at it then, by all means, by my guest!
Don't hesitate to ask, I'll gladly bring you up to speed with how it works. Following improvements: Growl
support (Mark.... where are you???), probably ImageMagick, hopefully a better set of build instructions
(specially with regards to TLS!)...

Thanks to all who helped: Markuss and Olegb for their reviews and suggestions, and of course to Jerome
Gagnon-Voyer (http://www.cmq.qc.ca/4w/amsn/) for readily providing the original instructions to build
the .app bundle out of the aMSN/TCL/TK sources, thanks! Now, enjoy... :-)


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@9436 d073be05-634f-4543-b044-5fe20cf6d1d6