Version bump to 3.0. Note that the directory hierarchy has changed.
Texi2html is no longer built as part of the teTeX port.  Instead, the
texi2html port has been made a dependency.  This avoids conflicts
in which the teTeX and texi2html ports put files in the same places.

As noted in the Portfile, the build is very fragile and is prone
to silent failures.  If you change the Portfile, make sure that all
of the targets you expect to be built really are. Check especially
that all of the expected symlinks get made.

Thanks to Richard Bonichon for the original Portfile changes!

Bug:
Submitted by:
Reviewed by:
Approved by:
Obtained from:


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@11011 d073be05-634f-4543-b044-5fe20cf6d1d6