Version bump to 0.41.
The patch file fixes a compilation bug caused by improper use of
C macros in a C++ program. (Not that the C++ spec is very clear on this
point, but the cognoscenti seem to understand.)  The bug is fixed in
the cvs version as part of a much more extensive cleanup. With luck,
the patch file will go away in the next release.

Bug:
Submitted by:
Reviewed by:
Approved by:
Obtained from:


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@11136 d073be05-634f-4543-b044-5fe20cf6d1d6