As discussed mith mww@, change the name of the shared library from "libpython.2.4.1.dylib" to "libpython2.4.1.dylib". The latter usage is more standard and is what other ports expect.
The additional build target still has to be called
libpython.2.4.1.dylib, so the fixes are all done in the post-destroot
phase (go figure).

This fixes bug 4935, where graphviz failed to build because it detected
python during configuration but could find the library to link.

Thanks to mww@, blb@ and Kyle Wheeler for the original report.
Bug:            4935
Submitted by:
Reviewed by:
Approved by:
Obtained from:


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@14076 d073be05-634f-4543-b044-5fe20cf6d1d6