Submitted by:   jmpp@
Adding dp@od as a second port maintainer for cyrus-imapd as I currently lack
any time what-so-ever to dedicate to this rather complex port, so that anyone
feeling brave enough to tackle it feels free to step up!

The port is currently out of date (which, among other things, implies adapting
the patch files) and in need of some lucky Intel Mac owner to give it some love.
If you, reading this, are one, do consider this as addressed to you! ;-)


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@18012 d073be05-634f-4543-b044-5fe20cf6d1d6