Change dependency on port:docbook-xml-412 to port:docbook-xml-4.1.2 Change file delete -force ... commands to delete ... commands Comment out delete ... commands that have nothing to delete in post-destroot Add sourceforge mastersites Change download from .bz2 to .gz since sourceforge download site only has .gz Bump revision so that no installed ports now rely on port:docbook-xml-412
NOTE: scrollkeeper-update runs in the destroot phase and crashes since it is looking for itself in ${prefix} not ${destroot}${prefix} at that time, but has not yet been installed. Once this is fixed, the post-destroot delete ... commands will need to be uncommented as we don't want any scrollkeeper data catalogs to be associated with any one port.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@25756 d073be05-634f-4543-b044-5fe20cf6d1d6