Change port from a specific docbook-xml DTD version to a meta-port that ensures that the user has all of the docbook-xml DTDs that MacPorts ships on his system.
Add a readme file that explains this and ensures that this port can be
registered as installed and is therefore upgradable.

Set the version on the port to the version of the highest versioned docbook-xml
DTD port.

This is the last change to the docbook-xml ports from
docbook-xml[|-scrollkeeper|-412] to docbook-xml-${version} before we can edit
ports that depended on the old docbook-xml ports to the new docbook-xml ports.

MAINTAINERS: Please use only the actual DTD versions that you need when setting
dependencies on your ports instead of depending on this port.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@25793 d073be05-634f-4543-b044-5fe20cf6d1d6