Commit patch contributed by Ryan in #12279 to tempoarily fix archival issue of packages that list two variant names in a single code block. Closes #12279.
NOTE: Consider this only a temporary, Portfile specific fix, the real bug in MacPorts' base should be
sought and fixed seperately. Tracked in #ticket 12306. As for this particular Portfile, Ryan's fix is
also a better approach 'cause 'darwin' and 'freebsd' shouldn't be listed as variants but platforms rather,
whereas 'kerberos' is a valid variant name.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@27045 d073be05-634f-4543-b044-5fe20cf6d1d6