Charlie is no longer with us, so I'm feeling bold enough to commit to this Portfile!
Move the menucolors variant to the end so that it is the last one for which its patchfile is fetched.
mirrors are lists tried in sequence, so if variant foo adds an entry, http://foo.com, for a file that's
only existent in it, all prior entries in the mirrors list are also tried for the file, naturally failing
the fetch. Good base enhancement there!

Also comment out the other variants, for which I updated their mirror sites but patchfiles are failing to checksum.
I do not have the necessary acquaintance with them to validate them, though.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@31957 d073be05-634f-4543-b044-5fe20cf6d1d6