deluge:
Remove the dependency on boost but error out in preconfigure check if boost +python25
is not installed with instructions to the user to install/reinstall the port manually
with the proper variant. This avoids the obnoxious and user unfriendly behavior of
installing boost without python25 support and then telling the user to dectivate it
and install boost +python25. Error message is made more prominent and hopefully easier
to understand for the naive user.

Partially addresses issues expressed in #126, #20057.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@52855 d073be05-634f-4543-b044-5fe20cf6d1d6