Fix incomplete build dependencies noted in #26762
A slight change in variant semantics: the default variant '+html' now
only installs the HTML and Info files.  The new variant '+pdf' now
installs documentation in HTML, PDF, and Info formats while properly
declaring dependency on the 'dvips' binary.  The revision remains the
same, as we do not wish to cause a reinstallation for users that have
successfully installed the port even though we have changed the
meaning of the variants.

Port now listed as openmaintainer as I have not heard a peep out of
waqar and gwright in over a year.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@73024 d073be05-634f-4543-b044-5fe20cf6d1d6