imake: Don't force llvm-gcc
imake itself builds fine with clang.  The problems experienced are
in ports that make use of imake, so the fixes or workarounds should
go into *those* ports, or the root cause should be addressed.

This just "worked" by setting imake's default compiler to something
different, so undoing that in the interest of fixing the real
problems.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@89535 d073be05-634f-4543-b044-5fe20cf6d1d6