{clang,llvm}-*: Additional fixes for Tiger and Leopard
Now that apple-gcc42 builds on Tiger, we can use it to build clang on Tiger.
This should allow us to now suport +universal on Tiger.

We also now use apple-gcc42 on Leopard because g++-4.0 produces some C++ that    
may not link, and the older host gcc-4.2 has some issues with trunk, so just
use the newer apple-gcc42 on those older platforms.

llvm-3.1 does not currently build ppc
clang-3.1 still does not build on Leopard and Tiger, but this brings us closer.


git-svn-id: http://svn.macports.org/repository/macports/trunk/dports@91469 d073be05-634f-4543-b044-5fe20cf6d1d6